+++ src/apis/Validator/index.tsx
@@ -1,6 +1,48 @@
-import { string } from 'prop-types';
 import { validatorAxios } from '../../instance/Axios';
-import { IValidator } from '../../instance/CryptoLabHandler';
+
+export interface IStatusChange {
+  commissionChange: number
+}
+
+export interface IIdentity {
+  display: string
+}
+
+export interface INominator {
+  address: string
+  balance: number
+}
+
+export interface IExposureOthers {
+  who: string
+  value: number
+}
+
+export interface IExposure {
+  total: number
+  own: number
+  others: IExposureOthers[]
+}
+
+export interface IEraInfo {
+  nominators: INominator[]
+  nominatorCount: number
+  era: number
+  exposure: IExposure
+  commission: number
+  apy: number
+  unclaimedEras: number[]
+  total: number
+}
+
+export interface IValidator {
+  id: string
+  statusChange: IStatusChange
+  identity: IIdentity
+  info: IEraInfo
+  averageApy: number
+  favorite: boolean
+}
 
 export interface IValidatorQuery {
   size?: number;
@@ -17,6 +59,6 @@ export interface IValidatorRequest {
   query?: IValidatorQuery;
 }
 export const apiGetAllValidator = (data: IValidatorRequest): Promise<IValidator[]> =>
-  validatorAxios.get(`${data.params}`, { params: data.query }).then((res) => {
-    return res.data;
-  });
+validatorAxios.get(`${data.params}`, { params: data.query }).then((res) => {
+  return res.data;
+});
