+++ src/instance/Axios.tsx
@@ -4,7 +4,7 @@ import keys from '../config/keys';
 export default Axios.create({});
 
 export const validatorAxios = Axios.create({
-  baseURL: `${keys.apiVersion}validators`,
+  baseURL: `${keys.apiVersion}/api/v1/validators`,
 });
 
 export const nominatorAxios = Axios.create({
